// src/app/model/usuarios/usuario-registrado-response.ts
export interface IUsuarioRegistradoResponse {
    nombreCompleto: string; // mapeado a nombre_completo
    apellidoPaterno: string; // mapeado a apellido_paterno
    apellidoMaterno: string; // mapeado a apellido_materno
    dni: string; // mapeado a dni
    correoElectronico: string; // mapeado a correo_electronico
    rol: string; // mapeado a rol
    telefono: string; // mapeado a telefono
}